Create core module
